requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.3
django-rest==0.8.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
timedelta==2020.12.3
uritemplate==4.1.1
